Refactor move reservability utils
